answerQuiz
Serg:
The onload now stores an array called questionList which conatins a map for each question - it has all the data in it including the randiomly selected levels  frio each question which will now stay fixed until the quiz is submitted. The images in the map are stored as an array which will only have 1 image unless its a multi image question, then there qwill be multiple.