Fixed sidebar issue on dashboards
There was an issue where if someone had a device with a screen small enough in height, the sidebar would only ever be the size of their screen initially. This meant that buttons overflowed so if the person zoomed out, the sidebar would not reach all the way to the bottom of the screen.